backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic==2.6.4
pydantic-settings==2.2.1
PyYAML==6.0.1
requests
firebase-admin
boto3
openai
moviepy==1.0.3
pillow
python-multipart

bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]
pydantic[email]

